Changed Homepage, Searchpage, and added FrameController
Created FrameController so we can switch between Views without creating
a multiple Frames (Which resulted in the window changing each time we
created a new frame). Also Added a JComponentWithLayout class just for
the FrameController to use. Run the program from the FrameController
now. Search works and the back button sends you back to the Homepage.
Will be able to add more columns when the MySQL is up and running or if
we actually implement it. Text me if you have any questions, I know I
changed a lot.